+++ crates/kobold/src/list/tracking.rs
